requirements.txt
numpy==1.19.1
Pillow==7.2.0
tensorflow==2.2.0
streamlit==0.63.1